Python-database-course/requirements.txt
alembic==1.10.4
APIFlask==0.12.0
apispec==5.1.1
appnope==0.1.3
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
boto3==1.28.17
botocore==1.31.51
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cors==1.0.1
cryptography==37.0.2
decorator==5.1.1
distlib==0.3.4
dnspython==2.2.1
docutils==0.16
email-validator==1.1.3
executing==0.8.3
Faker==23.2.1
filelock==3.6.0
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-DebugToolbar==0.13.1
Flask-HTTPAuth==4.5.0
Flask-JWT-Extended==4.4.0
flask-marshmallow==0.14.0
Flask-WTF==1.0.1
future==0.18.2
gevent==23.7.0
greenlet==2.0.2
gunicorn==20.0.4
idna==3.3
importlib-metadata==4.11.4
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
Mako==1.2.4
Markdown==3.3.7
MarkupSafe==2.1.1
marshmallow==3.15.0
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.4.0
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0.1
requests==2.27.1
requests-file==1.5.1
rsa==4.7.2
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.21
stack-data==0.2.0
stevedore==3.5.0
tldextract==3.3.0
toml==0.10.2
traitlets==5.1.1
typing_extensions==4.5.0
urllib3==1.26.9
virtualenv==20.13.4
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.5
webargs==8.1.0
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
